remove the copy_pipe_to loop optimization
This can take an arbitrary amount of time to finish if the output
throughput is lower than the input throughput. This would ideally be
fully non-blocking but the only way to achieve that on Linux is to
handle each standard stream in a separate thread with either a
non-blocking pipe or a more efficient eventfd design without any extra
copies.

Closes #50