rework the user interface for learning mode
Extending an existing system call whitelist is meant to be done with
that system call whitelist loaded. This avoids a performance hit for
system calls that are already whitelisted and will make it much easier
to add system call parameter support to the learning mode. Otherwise, it
would need to understand arbitrary parameter checks to avoid inserting
duplicate rules.